make all id long

+++ src/main/java/dao/BoxDAO.java
@@ -7,6 +7,6 @@ public interface BoxDAO {
 	public boolean operatonWBox(Box box, Instrument instrument,float amount,boolean type);
 	public Box getBoxByNumber(int number,int idLocation);
 	public boolean createBox(Box box);
-	public boolean deleteBox(int id);
+	public boolean deleteBox(long id);
 	public boolean hasError();
 }
