+++ src/main/java/dao/LocationDAO.java
@@ -6,9 +6,9 @@ import models.Location;
 
 public interface LocationDAO {
 	public  boolean createLocation(Location location);
-	public Location getLocById(int id);
+	public Location getLocById(long id);
 	public Location getLocByName(String name);
 	public List<Location> getLocByNameL(String name);
-	public boolean deleteLocation(int id);
+	public boolean deleteLocation(long id);
 	public boolean hasError();
 }
