+++ src/main/java/dao/OutDocDAO.java
@@ -11,6 +11,6 @@ public interface OutDocDAO {
 	public List<OutDoc> getOutDocByDate(LocalDate date);
 	public List<OutDoc> getOutDocByInstrum(long id);
 	public List<OutDoc> getOutDocByBox(long idB,long idL);
-	public boolean deleteInDoc(int id);
+	public boolean deleteInDoc(long id);
 	public boolean hasError();
 }
